+++ crates/duckdb/CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `Client.search_to_arrow_table` ([#634](https://github.com/stac-utils/stac-rs/pull/634))
 - Conditionally disable parsing the WKB ([#635](https://github.com/stac-utils/stac-rs/pull/635))
 
+### Removed
+
+- geoarrow record batch converters (moved to **stac**) ([#652](https://github.com/stac-utils/stac-rs/pull/652))
+
 ## [0.1.1] - 2025-01-31
 
 ### Added